feat: swap tip
