+++ packages/kit-bg/src/vaults/impls/btc/Vault.ts
@@ -250,6 +250,7 @@ export default class VaultBtc extends VaultBase {
             amount: swapInfo.sender.amount,
             isNFT: false,
             isNative: swapSendToken.isNative,
+            networkId: swapInfo.sender.accountInfo.networkId,
           },
           {
             from: '',
@@ -261,6 +262,7 @@ export default class VaultBtc extends VaultBase {
             amount: swapInfo.receiver.amount,
             isNFT: false,
             isNative: swapReceiveToken.isNative,
+            networkId: swapInfo.receiver.accountInfo.networkId,
           },
         ],
         isInternalSwap: true,
