+++ packages/kit-bg/src/vaults/impls/kaspa/Vault.ts
@@ -208,6 +208,7 @@ export default class Vault extends VaultBase {
             amount: swapInfo.sender.amount,
             isNFT: false,
             isNative: swapSendToken.isNative,
+            networkId: swapInfo.sender.accountInfo.networkId,
           },
           {
             from: '',
@@ -219,6 +220,7 @@ export default class Vault extends VaultBase {
             amount: swapInfo.receiver.amount,
             isNFT: false,
             isNative: swapReceiveToken.isNative,
+            networkId: swapInfo.receiver.accountInfo.networkId,
           },
         ],
         isInternalSwap: true,
