+++ packages/kit/src/views/Swap/components/SwapProviderInfoItem.tsx
@@ -1,22 +1,20 @@
-import { memo, useCallback, useMemo } from 'react';
+import { memo } from 'react';
 
 import { useIntl } from 'react-intl';
 
-import type { IImageSourceProps } from '@onekeyhq/components';
 import {
   Badge,
   Icon,
-  IconButton,
   Image,
-  Popover,
   SizableText,
   Skeleton,
   Stack,
   XStack,
 } from '@onekeyhq/components';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
-import { otherWalletFeeData } from '@onekeyhq/shared/types/swap/SwapProvider.constants';
 import type { ISwapToken } from '@onekeyhq/shared/types/swap/types';
+
+import { SwapServiceFeeOverview } from './SwapServiceFeeOverview';
 
 interface ISwapProviderInfoItemProps {
   fromToken?: ISwapToken;
@@ -29,13 +27,6 @@ interface ISwapProviderInfoItemProps {
   onPress?: () => void;
   isLoading?: boolean;
 }
-interface IProtocolFeeInfo {
-  name: string;
-  fee: number;
-  color: string;
-  icon: IImageSourceProps['source'];
-  maxFee: number;
-}
 
 const SwapProviderInfoItemTitleContent = ({
   onekeyFee,
@@ -43,45 +34,7 @@ const SwapProviderInfoItemTitleContent = ({
   onekeyFee?: number;
 }) => {
   const intl = useIntl();
-  const protocolFeeInfoList: IProtocolFeeInfo[] = useMemo(
-    () => [
-      ...otherWalletFeeData,
-      {
-        maxFee: 0.875,
-        name: 'oneKey',
-        fee: onekeyFee ?? 0.3,
-        // color: '#202020',
-        color: '$bgInverse',
-        icon: require('@onekeyhq/kit/assets/logo.png'),
-      },
-    ],
-    [onekeyFee],
-  );
-  const renderProtocolFeeListItem = useCallback(
-    (item: IProtocolFeeInfo) => (
-      <XStack gap="$3" alignItems="center">
-        <Stack w={20} h={20}>
-          <Image source={item.icon} w={16} h={16} />
-        </Stack>
-        <Stack flex={1}>
-          <Stack
-            backgroundColor={item.color}
-            borderRadius="$full"
-            width={`${item.maxFee > 0 ? (item.fee / item.maxFee) * 100 : 0}%`}
-            height="$1"
-          />
-        </Stack>
-        <SizableText
-          size="$bodySm"
-          color={item.name === 'oneKey' ? '$textSuccess' : '$text'}
-          textAlign="right"
-        >
-          {item.fee}%
-        </SizableText>
-      </XStack>
-    ),
-    [],
-  );
+
   return (
     <XStack>
       <SizableText
@@ -94,37 +47,12 @@ const SwapProviderInfoItemTitleContent = ({
           id: ETranslations.swap_page_provider_provider,
         })}
       </SizableText>
-      <Popover
-        title={intl.formatMessage({
-          id: ETranslations.provider_ios_popover_onekey_fee,
-        })}
-        renderTrigger={
-          <IconButton
-            variant="tertiary"
-            size="small"
-            icon="InfoCircleOutline"
-          />
-        }
-        renderContent={
-          <Stack gap="$5" p="$4">
-            <SizableText size="$bodyMd" color="$textSubdued">
-              {intl.formatMessage({
-                id: ETranslations.provider_ios_popover_onekey_fee_content,
-              })}
-            </SizableText>
-            <Stack gap="$2">
-              {protocolFeeInfoList.map((item) =>
-                renderProtocolFeeListItem(item),
-              )}
-            </Stack>
-          </Stack>
-        }
-      />
+      <SwapServiceFeeOverview onekeyFee={onekeyFee} />
     </XStack>
   );
 };
 
-const SwapProviderInfoItemTitleContentMemo = memo(
+export const SwapProviderInfoItemTitleContentMemo = memo(
   SwapProviderInfoItemTitleContent,
 );
