+++ packages/shared/types/tx.ts
@@ -158,6 +158,7 @@ export type IDecodedTxTransferInfo = {
   isOwn?: boolean; // for UTXO
   label?: string;
   price?: string;
+  networkId?: string;
 };
 
 export type IDecodedTxActionFunctionCall = IDecodedTxActionBase & {
